Refactored openfin from blotter into openfinEpics
