+++ src/client/src/openfinEpics.ts
@@ -1,11 +1,28 @@
-import { combineEpics, ofType } from 'redux-observable'
+import { combineEpics, ofType, StateObservable } from 'redux-observable'
 import { bindCallback } from 'rxjs'
 import { filter, map, mergeMap, switchMapTo, takeUntil, tap, withLatestFrom } from 'rxjs/operators'
 import { ApplicationEpic } from './ApplicationEpic'
+import { GlobalState } from './combineReducers'
 import { CONNECT_SERVICES, DISCONNECT_SERVICES } from './connectionActions'
-import { Direction, SpotPriceTick } from './types'
+import { CurrencyPairReducerState } from './currencyPairsOperations'
+import { OpenFin } from './services'
+import { Direction, SpotPriceTick, Trades } from './types'
 import { CurrencyPair } from './types/currencyPair'
+import { ACTION_TYPES as BLOTTER_ACTION_TYPES } from './ui/blotter/actions'
 import { SpotTileActions } from './ui/spotTile/actions'
+
+const subscribeOpenFinToBlotterData = (openFin: OpenFin, state$: StateObservable<GlobalState>) => () => {
+  const trades: Trades = state$.value.blotterService.trades
+  const currencyPairs: CurrencyPairReducerState = state$.value.currencyPairs
+  const cb = (msg: any, uuid: string) => openFin.sendAllBlotterData(uuid, trades, currencyPairs)
+  openFin.addSubscription('fetch-blotter', cb)
+}
+
+export const connectBlotterServiceToOpenFinEpic: ApplicationEpic = (action$, state$, { openFin }) =>
+  action$.pipe(
+    ofType(BLOTTER_ACTION_TYPES.BLOTTER_SERVICE_NEW_TRADES),
+    tap(subscribeOpenFinToBlotterData(openFin, state$))
+  )
 
 const addRatePrecisionToPrice = (currencyData: Map<string, CurrencyPair>, price: SpotPriceTick) => {
   return {
@@ -66,4 +83,4 @@ export const closePositionEpic: ApplicationEpic = (action$, state$, { openFin })
     })
   )
 
-export const openfinEpic = combineEpics(publishPriceToOpenFinEpic)
+export const openfinEpic = combineEpics(publishPriceToOpenFinEpic, connectBlotterServiceToOpenFinEpic)
